CombinedVolumeContexts.
