+++ src/components/sounds/AverageSound.jsx
@@ -1,12 +1,12 @@
 import React, { useEffect } from "react";
-import { useVolume } from "../../contexts/VolumeContext";
+import { useVolumeSettings } from "../../contexts/VolumeContext";
 import sound1 from "./SoundAssets/correct-check.mp3"; // Import the audio file
 import sound2 from "./SoundAssets/ding-effect.mp3"; // Import the audio file
 import sound3 from "./SoundAssets/correct-answer-sound.mp3"; // Import the audio file
 
 
 function App() {
-  const { volume } = useVolume(); // Access volume state from VolumeContext
+  const { volume } = useVolumeSettings(); // Access volume state from VolumeContext
 
   useEffect(() => {
     const sounds = [sound1, sound2, sound3];
